Publisher example: Extract PublisherInfoViewModel
I’m going to be moving this information out of the map view.
